Added python scripts that facilitate various tasks.
CheckApi.py parses core/src/fieldml_api.h and io/src/FieldmlIoApi.h and compares the function declarations in those files with non-static function definitions in the corresponding .cpp files to make sure that every declared function is defined, and vice versa. It also ensures that parameter names are the same. This is used to consistency-check the API headers and sources against each other.

FileToString.py converts a file into a c-style string literal by escaping quotes and inserting linefeeds. This is currently only used to generate the built-in XSD and library strings.

ExtractStrings.py extracts element and attribute names from a given XSD file and generates either a set of const char* declarations or definitions, depending on a boolean runtime argument. This is used to quickly generate the relevant consts from an edited XSD file for use with the XML I/O code.